refactor flags + add tlsInsecure flag + change default TLS config
