+++ cli/flags/bind_healthcheck_flags.go
@@ -1,0 +1,9 @@
+package flags
+
+import (
+	"github.com/spf13/cobra"
+)
+
+func SetHealthCheckFlags(cmd *cobra.Command) {
+	AddPersistentStringSliceFlag(cmd, "ip", []string{}, "Operator ip:port", true)
+}
