+++ cli/initiator/reshare.go
@@ -10,6 +10,7 @@ import (
 	"go.uber.org/zap"
 
 	spec "github.com/ssvlabs/dkg-spec"
+	"github.com/ssvlabs/ssv-dkg/cli/flags"
 	cli_utils "github.com/ssvlabs/ssv-dkg/cli/utils"
 	"github.com/ssvlabs/ssv-dkg/pkgs/initiator"
 	"github.com/ssvlabs/ssv-dkg/pkgs/utils"
@@ -17,21 +18,21 @@ import (
 )
 
 func init() {
-	cli_utils.SetGenerateReshareMsgFlags(GenerateReshareMsg)
-	cli_utils.SetReshareFlags(StartReshare)
+	flags.SetGenerateReshareMsgFlags(GenerateReshareMsg)
+	flags.SetReshareFlags(StartReshare)
 }
 
 var GenerateReshareMsg = &cobra.Command{
 	Use:   "generate-reshare-msg",
 	Short: "Generate reshare message hash for one or multiple ceremonies",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		if err := cli_utils.SetViperConfig(cmd); err != nil {
+		if err := flags.SetViperConfig(cmd); err != nil {
 			return err
 		}
-		if err := cli_utils.BindGenerateReshareMsgFlags(cmd); err != nil {
+		if err := flags.BindGenerateReshareMsgFlags(cmd); err != nil {
 			return err
 		}
-		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator")
+		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator", flags.LogFilePath, flags.LogLevel, flags.LogFormat, flags.LogLevelFormat)
 		if err != nil {
 			return err
 		}
@@ -41,53 +42,53 @@ var GenerateReshareMsg = &cobra.Command{
 			}
 		}()
 		logger.Info("🪛 Initiator`s", zap.String("Version", cmd.Version))
-		opMap, err := cli_utils.LoadOperators(logger)
+		opMap, err := cli_utils.LoadOperators(logger, flags.OperatorsInfo, flags.OperatorsInfoPath)
 		if err != nil {
 			logger.Fatal("😥 Failed to load operators: ", zap.Error(err))
 		}
-		oldOperatorIDs, err := cli_utils.StringSliceToUintArray(cli_utils.OperatorIDs)
+		oldOperatorIDs, err := cli_utils.StringSliceToUintArray(flags.OperatorIDs)
 		if err != nil {
 			logger.Fatal("😥 Failed to load participants: ", zap.Error(err))
 		}
-		newOperatorIDs, err := cli_utils.StringSliceToUintArray(cli_utils.NewOperatorIDs)
+		newOperatorIDs, err := cli_utils.StringSliceToUintArray(flags.NewOperatorIDs)
 		if err != nil {
 			logger.Fatal("😥 Failed to load new participants: ", zap.Error(err))
 		}
 		// create initiator instance
-		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, cli_utils.ClientCACertPath)
+		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, nil, true)
 		if err != nil {
 			return err
 		}
 		var signedProofs [][]*spec.SignedProof
-		if cli_utils.ProofsFilePath != "" {
-			signedProofs, err = wire.LoadProofs(cli_utils.ProofsFilePath)
+		if flags.ProofsFilePath != "" {
+			signedProofs, err = wire.LoadProofs(flags.ProofsFilePath)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs json file:", zap.Error(err))
 			}
 		}
-		if cli_utils.ProofsString != "" {
-			signedProofs, err = cli_utils.DecodeProofsString(cli_utils.ProofsString)
+		if flags.ProofsString != "" {
+			signedProofs, err = cli_utils.DecodeProofsString(flags.ProofsString)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs string:", zap.Error(err))
 			}
 		}
-		ethNetwork := e2m_core.NetworkFromString(cli_utils.Network)
+		ethNetwork := e2m_core.NetworkFromString(flags.Network)
 		if ethNetwork == "" {
 			logger.Fatal("😥 Cant recognize eth network")
 		}
 		rMsgs := []*wire.ReshareMessage{}
 		for i := 0; i < len(signedProofs); i++ {
-			nonce := cli_utils.Nonce + uint64(i)
-			// Contruct the resign message
+			nonce := flags.Nonce + uint64(i)
+			// Construct reshare message
 			rMsg, err := dkgInitiator.ConstructReshareMessage(
 				oldOperatorIDs,
 				newOperatorIDs,
 				signedProofs[i][0].Proof.ValidatorPubKey,
 				ethNetwork,
-				cli_utils.WithdrawAddress[:],
-				cli_utils.OwnerAddress,
+				flags.WithdrawAddress[:],
+				flags.OwnerAddress,
 				nonce,
-				cli_utils.Amount,
+				flags.Amount,
 				signedProofs[i],
 			)
 			if err != nil {
@@ -100,7 +101,7 @@ var GenerateReshareMsg = &cobra.Command{
 		if err != nil {
 			logger.Fatal("😥 Failed to marshal reshare message hash:", zap.Error(err))
 		}
-		finalPath := fmt.Sprintf("%s/reshare.txt", cli_utils.OutputPath)
+		finalPath := fmt.Sprintf("%s/reshare.txt", flags.OutputPath)
 		err = os.WriteFile(finalPath, []byte(msgHex), 0o600)
 		if err != nil {
 			logger.Fatal("😥 Failed to save reshare message hash:", zap.Error(err))
@@ -125,13 +126,13 @@ var StartReshare = &cobra.Command{
 		░ ░  ░ ░ ░░ ░ ░ ░   ░      ░░   ░    ░   ░  ░  ░   ░  ░░ ░  ░   ▒     ░░   ░    ░   
 		░    ░  ░         ░       ░        ░  ░      ░   ░  ░  ░      ░  ░   ░        ░  ░
 		░`)
-		if err := cli_utils.SetViperConfig(cmd); err != nil {
+		if err := flags.SetViperConfig(cmd); err != nil {
 			return err
 		}
-		if err := cli_utils.BindReshareFlags(cmd); err != nil {
+		if err := flags.BindReshareFlags(cmd); err != nil {
 			return err
 		}
-		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator")
+		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator", flags.LogFilePath, flags.LogLevel, flags.LogFormat, flags.LogLevelFormat)
 		if err != nil {
 			return err
 		}
@@ -141,59 +142,59 @@ var StartReshare = &cobra.Command{
 			}
 		}()
 		logger.Info("🪛 Initiator`s", zap.String("Version", cmd.Version))
-		opMap, err := cli_utils.LoadOperators(logger)
+		opMap, err := cli_utils.LoadOperators(logger, flags.OperatorsInfo, flags.OperatorsInfoPath)
 		if err != nil {
 			logger.Fatal("😥 Failed to load operators: ", zap.Error(err))
 		}
-		oldOperatorIDs, err := cli_utils.StringSliceToUintArray(cli_utils.OperatorIDs)
+		oldOperatorIDs, err := cli_utils.StringSliceToUintArray(flags.OperatorIDs)
 		if err != nil {
 			logger.Fatal("😥 Failed to load participants: ", zap.Error(err))
 		}
-		newOperatorIDs, err := cli_utils.StringSliceToUintArray(cli_utils.NewOperatorIDs)
+		newOperatorIDs, err := cli_utils.StringSliceToUintArray(flags.NewOperatorIDs)
 		if err != nil {
 			logger.Fatal("😥 Failed to load new participants: ", zap.Error(err))
 		}
 		// create a new ID for resharing
 		id := spec.NewID()
 		// create initiator instance
-		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, cli_utils.ClientCACertPath)
+		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, flags.ClientCACertPath, flags.TLSInsecure)
 		if err != nil {
 			return err
 		}
 		var signedProofs [][]*spec.SignedProof
-		if cli_utils.ProofsFilePath != "" {
-			signedProofs, err = wire.LoadProofs(cli_utils.ProofsFilePath)
+		if flags.ProofsFilePath != "" {
+			signedProofs, err = wire.LoadProofs(flags.ProofsFilePath)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs json file:", zap.Error(err))
 			}
 		}
-		if cli_utils.ProofsString != "" {
-			signedProofs, err = cli_utils.DecodeProofsString(cli_utils.ProofsString)
+		if flags.ProofsString != "" {
+			signedProofs, err = cli_utils.DecodeProofsString(flags.ProofsString)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs string:", zap.Error(err))
 			}
 		}
-		ethNetwork := e2m_core.NetworkFromString(cli_utils.Network)
+		ethNetwork := e2m_core.NetworkFromString(flags.Network)
 		if ethNetwork == "" {
 			logger.Fatal("😥 Cant recognize eth network")
 		}
-		signatures, err := cli_utils.SignaturesStringToBytes(cli_utils.Signatures)
+		signatures, err := cli_utils.SignaturesStringToBytes(flags.Signatures)
 		if err != nil {
 			logger.Fatal("😥 Failed to load signatures: ", zap.Error(err))
 		}
 		rMsgs := []*wire.ReshareMessage{}
 		for i := 0; i < len(signedProofs); i++ {
-			nonce := cli_utils.Nonce + uint64(i)
+			nonce := flags.Nonce + uint64(i)
 			// Contruct the reshare message
 			rMsg, err := dkgInitiator.ConstructReshareMessage(
 				oldOperatorIDs,
 				newOperatorIDs,
 				signedProofs[i][0].Proof.ValidatorPubKey,
 				ethNetwork,
-				cli_utils.WithdrawAddress[:],
-				cli_utils.OwnerAddress,
+				flags.WithdrawAddress[:],
+				flags.OwnerAddress,
 				nonce,
-				cli_utils.Amount,
+				flags.Amount,
 				signedProofs[i],
 			)
 			if err != nil {
@@ -220,10 +221,10 @@ var StartReshare = &cobra.Command{
 			proofs,
 			false,
 			len(signedProofs),
-			cli_utils.OwnerAddress,
-			cli_utils.Nonce,
-			cli_utils.WithdrawAddress,
-			cli_utils.OutputPath,
+			flags.OwnerAddress,
+			flags.Nonce,
+			flags.WithdrawAddress,
+			flags.OutputPath,
 		); err != nil {
 			logger.Fatal("Could not save results", zap.Error(err))
 		}
