+++ cli/initiator/resigning.go
@@ -10,6 +10,7 @@ import (
 	"go.uber.org/zap"
 
 	spec "github.com/ssvlabs/dkg-spec"
+	"github.com/ssvlabs/ssv-dkg/cli/flags"
 	cli_utils "github.com/ssvlabs/ssv-dkg/cli/utils"
 	"github.com/ssvlabs/ssv-dkg/pkgs/initiator"
 	"github.com/ssvlabs/ssv-dkg/pkgs/utils"
@@ -17,21 +18,21 @@ import (
 )
 
 func init() {
-	cli_utils.SetGenerateResignMsgFlags(GenerateResignMsg)
-	cli_utils.SetResigningFlags(StartResigning)
+	flags.SetGenerateResignMsgFlags(GenerateResignMsg)
+	flags.SetResigningFlags(StartResigning)
 }
 
 var GenerateResignMsg = &cobra.Command{
 	Use:   "generate-resign-msg",
 	Short: "Generate resign message hash for one or multiple ceremonies",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		if err := cli_utils.SetViperConfig(cmd); err != nil {
+		if err := flags.SetViperConfig(cmd); err != nil {
 			return err
 		}
-		if err := cli_utils.BindGenerateResignMsgFlags(cmd); err != nil {
+		if err := flags.BindGenerateResignMsgFlags(cmd); err != nil {
 			return err
 		}
-		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator")
+		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator", flags.LogFilePath, flags.LogLevel, flags.LogFormat, flags.LogLevelFormat)
 		if err != nil {
 			return err
 		}
@@ -41,48 +42,48 @@ var GenerateResignMsg = &cobra.Command{
 			}
 		}()
 		logger.Info("🪛 Initiator`s", zap.String("Version", cmd.Version))
-		opMap, err := cli_utils.LoadOperators(logger)
+		opMap, err := cli_utils.LoadOperators(logger, flags.OperatorsInfo, flags.OperatorsInfoPath)
 		if err != nil {
 			logger.Fatal("😥 Failed to load operators: ", zap.Error(err))
 		}
-		operatorIDs, err := cli_utils.StringSliceToUintArray(cli_utils.OperatorIDs)
+		operatorIDs, err := cli_utils.StringSliceToUintArray(flags.OperatorIDs)
 		if err != nil {
 			logger.Fatal("😥 Failed to load participants: ", zap.Error(err))
 		}
-		ethNetwork := e2m_core.NetworkFromString(cli_utils.Network)
+		ethNetwork := e2m_core.NetworkFromString(flags.Network)
 		if ethNetwork == "" {
 			logger.Fatal("😥 Cant recognize eth network")
 		}
 		var signedProofs [][]*spec.SignedProof
-		if cli_utils.ProofsFilePath != "" {
-			signedProofs, err = wire.LoadProofs(cli_utils.ProofsFilePath)
+		if flags.ProofsFilePath != "" {
+			signedProofs, err = wire.LoadProofs(flags.ProofsFilePath)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs json file:", zap.Error(err))
 			}
 		}
-		if cli_utils.ProofsString != "" {
-			signedProofs, err = cli_utils.DecodeProofsString(cli_utils.ProofsString)
+		if flags.ProofsString != "" {
+			signedProofs, err = cli_utils.DecodeProofsString(flags.ProofsString)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs string:", zap.Error(err))
 			}
 		}
 		// Create new DKG initiator
-		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, cli_utils.ClientCACertPath)
+		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, nil, true)
 		if err != nil {
 			return err
 		}
 		// Reconstruct the resign messages
 		rMsgs := []*wire.ResignMessage{}
 		for i := 0; i < len(signedProofs); i++ {
-			nonce := cli_utils.Nonce + uint64(i)
+			nonce := flags.Nonce + uint64(i)
 			rMsg, err := dkgInitiator.ConstructResignMessage(
 				operatorIDs,
 				signedProofs[i][0].Proof.ValidatorPubKey,
 				ethNetwork,
-				cli_utils.WithdrawAddress[:],
-				cli_utils.OwnerAddress,
+				flags.WithdrawAddress[:],
+				flags.OwnerAddress,
 				nonce,
-				cli_utils.Amount,
+				flags.Amount,
 				signedProofs[i],
 			)
 			if err != nil {
@@ -95,7 +96,7 @@ var GenerateResignMsg = &cobra.Command{
 		if err != nil {
 			logger.Fatal("😥 Failed to marshal resign message hash:", zap.Error(err))
 		}
-		finalPath := fmt.Sprintf("%s/resign.txt", cli_utils.OutputPath)
+		finalPath := fmt.Sprintf("%s/resign.txt", flags.OutputPath)
 		err = os.WriteFile(finalPath, []byte(msgHex), 0o600)
 		if err != nil {
 			logger.Fatal("😥 Failed to save resign message hash:", zap.Error(err))
@@ -118,13 +119,13 @@ var StartResigning = &cobra.Command{
 		╚═════╝ ╚═╝  ╚═╝ ╚═════╝     ╚═╝  ╚═╝╚══════╝╚══════╝╚═╝ ╚═════╝ ╚═╝  ╚═══╝
 																				`)
 
-		if err := cli_utils.SetViperConfig(cmd); err != nil {
+		if err := flags.SetViperConfig(cmd); err != nil {
 			return err
 		}
-		if err := cli_utils.BindResigningFlags(cmd); err != nil {
+		if err := flags.BindResigningFlags(cmd); err != nil {
 			return err
 		}
-		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator")
+		logger, err := cli_utils.SetGlobalLogger(cmd, "dkg-initiator", flags.LogFilePath, flags.LogLevel, flags.LogFormat, flags.LogLevelFormat)
 		if err != nil {
 			return err
 		}
@@ -135,37 +136,37 @@ var StartResigning = &cobra.Command{
 		}()
 		logger.Info("🪛 Initiator`s", zap.String("Version", cmd.Version))
 		// Load operators
-		opMap, err := cli_utils.LoadOperators(logger)
+		opMap, err := cli_utils.LoadOperators(logger, flags.OperatorsInfo, flags.OperatorsInfoPath)
 		if err != nil {
 			logger.Fatal("😥 Failed to load operators: ", zap.Error(err))
 		}
-		operatorIDs, err := cli_utils.StringSliceToUintArray(cli_utils.OperatorIDs)
+		operatorIDs, err := cli_utils.StringSliceToUintArray(flags.OperatorIDs)
 		if err != nil {
 			logger.Fatal("😥 Failed to load participants: ", zap.Error(err))
 		}
-		ethNetwork := e2m_core.NetworkFromString(cli_utils.Network)
+		ethNetwork := e2m_core.NetworkFromString(flags.Network)
 		if ethNetwork == "" {
 			logger.Fatal("😥 Cant recognize eth network")
 		}
 		var signedProofs [][]*spec.SignedProof
-		if cli_utils.ProofsFilePath != "" {
-			signedProofs, err = wire.LoadProofs(cli_utils.ProofsFilePath)
+		if flags.ProofsFilePath != "" {
+			signedProofs, err = wire.LoadProofs(flags.ProofsFilePath)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs json file:", zap.Error(err))
 			}
 		}
-		if cli_utils.ProofsString != "" {
-			signedProofs, err = cli_utils.DecodeProofsString(cli_utils.ProofsString)
+		if flags.ProofsString != "" {
+			signedProofs, err = cli_utils.DecodeProofsString(flags.ProofsString)
 			if err != nil {
 				logger.Fatal("😥 Failed to read proofs string:", zap.Error(err))
 			}
 		}
-		signatures, err := cli_utils.SignaturesStringToBytes(cli_utils.Signatures)
+		signatures, err := cli_utils.SignaturesStringToBytes(flags.Signatures)
 		if err != nil {
 			logger.Fatal("😥 Failed to load signatures: ", zap.Error(err))
 		}
 		// Create new DKG initiator
-		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, cli_utils.ClientCACertPath)
+		dkgInitiator, err := initiator.New(opMap.Clone(), logger, cmd.Version, flags.ClientCACertPath, flags.TLSInsecure)
 		if err != nil {
 			return err
 		}
@@ -174,14 +175,14 @@ var StartResigning = &cobra.Command{
 		// Reconstruct the resign messages
 		rMsgs := []*wire.ResignMessage{}
 		for i := 0; i < len(signedProofs); i++ {
-			nonce := cli_utils.Nonce + uint64(i)
+			nonce := flags.Nonce + uint64(i)
 			rMsg, err := dkgInitiator.ConstructResignMessage(
 				operatorIDs,
 				signedProofs[i][0].Proof.ValidatorPubKey,
 				ethNetwork,
-				cli_utils.WithdrawAddress[:],
-				cli_utils.OwnerAddress,
-				nonce, cli_utils.Amount,
+				flags.WithdrawAddress[:],
+				flags.OwnerAddress,
+				nonce, flags.Amount,
 				signedProofs[i],
 			)
 			if err != nil {
@@ -208,10 +209,10 @@ var StartResigning = &cobra.Command{
 			proofs,
 			false,
 			len(signedProofs),
-			cli_utils.OwnerAddress,
-			cli_utils.Nonce,
-			cli_utils.WithdrawAddress,
-			cli_utils.OutputPath,
+			flags.OwnerAddress,
+			flags.Nonce,
+			flags.WithdrawAddress,
+			flags.OutputPath,
 		); err != nil {
 			logger.Fatal("Could not save results", zap.Error(err))
 		}
