+++ cli/verify/verify.go
@@ -5,31 +5,31 @@ import (
 	"log"
 	"os"
 
-	"github.com/spf13/cobra"
 	"github.com/aquasecurity/table"
+	"github.com/spf13/cobra"
 
-	cli_utils "github.com/ssvlabs/ssv-dkg/cli/utils"
+	"github.com/ssvlabs/ssv-dkg/cli/flags"
 	"github.com/ssvlabs/ssv-dkg/pkgs/validator"
 )
 
 func init() {
-	cli_utils.SetVerifyFlags(Verify)
+	flags.SetVerifyFlags(Verify)
 }
 
 var Verify = &cobra.Command{
 	Use:   "verify",
 	Short: "Verifies a DKG ceremony directory",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		if err := cli_utils.BindVerifyFlags(cmd); err != nil {
+		if err := flags.BindVerifyFlags(cmd); err != nil {
 			return err
 		}
 
 		err := validator.ValidateResultsDir(
-			cli_utils.CeremonyDir,
-			int(cli_utils.Validators),
-			cli_utils.OwnerAddress,
-			cli_utils.Nonce,
-			cli_utils.WithdrawAddress,
+			flags.CeremonyDir,
+			int(flags.Validators),
+			flags.OwnerAddress,
+			flags.Nonce,
+			flags.WithdrawAddress,
 		)
 		if err != nil {
 			log.Printf("Failed to validate ceremony directory: %v", err)
@@ -41,11 +41,11 @@ var Verify = &cobra.Command{
 		tbl := table.New(os.Stdout)
 		tbl.SetHeaders("Directory", "Withdrawal Address", "Nonce", "Owner Address", "Validators")
 		tbl.AddRow(
-			cli_utils.CeremonyDir,
-			cli_utils.WithdrawAddress.String(),
-			fmt.Sprintf("%d", cli_utils.Nonce),
-			cli_utils.OwnerAddress.String(),
-			fmt.Sprintf("%d", cli_utils.Validators),
+			flags.CeremonyDir,
+			flags.WithdrawAddress.String(),
+			fmt.Sprintf("%d", flags.Nonce),
+			flags.OwnerAddress.String(),
+			fmt.Sprintf("%d", flags.Validators),
 		)
 		tbl.Render()
