+++ integration_test/init_bulk_test.go
@@ -43,7 +43,15 @@ func TestBulkHappyFlows4Ops(t *testing.T) {
 	cli_initiator.StartDKG.Version = version
 	cli_verify.Verify.Version = version
 	t.Run("test 4 operators 1 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "1", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44", "--nonce", "1", "--amount", "32000000000",  "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "1",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -51,14 +59,30 @@ func TestBulkHappyFlows4Ops(t *testing.T) {
 	})
 
 	t.Run("test 4 operators 10 validators bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "10", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "10",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
 		resetFlags(RootCmd)
 	})
 	t.Run("test 4 operators 100 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "100", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "100",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -69,7 +93,13 @@ func TestBulkHappyFlows4Ops(t *testing.T) {
 	require.NoError(t, err)
 	validators := []int{1, 10, 100}
 	for i, c := range initCeremonies {
-		args := []string{"verify", "--ceremonyDir", "./output/" + c.Name(), "--validators", strconv.Itoa(validators[i]), "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--nonce", strconv.Itoa(1),  "--amount", "32000000000"}
+		args := []string{"verify",
+			"--ceremonyDir", "./output/" + c.Name(),
+			"--validators", strconv.Itoa(validators[i]),
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--nonce", strconv.Itoa(1),
+			"--amount", "32000000000"}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -109,21 +139,45 @@ func TestBulkHappyFlows7Ops(t *testing.T) {
 	cli_initiator.StartDKG.Version = version
 	cli_verify.Verify.Version = version
 	t.Run("test 7 operators 1 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "1", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "1",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
 		resetFlags(RootCmd)
 	})
 	t.Run("test 7 operators 10 validators bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "10", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "10",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
 		resetFlags(RootCmd)
 	})
 	t.Run("test 7 operators 100 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "100", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "100",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -134,7 +188,13 @@ func TestBulkHappyFlows7Ops(t *testing.T) {
 	require.NoError(t, err)
 	validators := []int{1, 10, 100}
 	for i, c := range initCeremonies {
-		args := []string{"verify", "--ceremonyDir", "./output/" + c.Name(), "--validators", strconv.Itoa(validators[i]), "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--nonce", strconv.Itoa(1), "--amount", "32000000000"}
+		args := []string{"verify",
+			"--ceremonyDir", "./output/" + c.Name(),
+			"--validators", strconv.Itoa(validators[i]),
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--nonce", strconv.Itoa(1),
+			"--amount", "32000000000"}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -174,21 +234,45 @@ func TestBulkHappyFlows10Ops(t *testing.T) {
 	cli_initiator.StartDKG.Version = version
 	cli_verify.Verify.Version = version
 	t.Run("test 10 operators 1 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "1", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77,88,99,100", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "1",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77,88,99,100",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
 		resetFlags(RootCmd)
 	})
 	t.Run("test 10 operators 10 validators bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "10", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77,88,99,100", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "10",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77,88,99,100",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
 		resetFlags(RootCmd)
 	})
 	t.Run("test 10 operators 100 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "100", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77,88,99,100", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "100",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77,88,99,100",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -199,7 +283,13 @@ func TestBulkHappyFlows10Ops(t *testing.T) {
 	require.NoError(t, err)
 	validators := []int{1, 10, 100}
 	for i, c := range initCeremonies {
-		args := []string{"verify", "--ceremonyDir", "./output/" + c.Name(), "--validators", strconv.Itoa(validators[i]), "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--nonce", strconv.Itoa(1), "--amount", "32000000000"}
+		args := []string{"verify",
+			"--ceremonyDir", "./output/" + c.Name(),
+			"--validators", strconv.Itoa(validators[i]),
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--nonce", strconv.Itoa(1),
+			"--amount", "32000000000"}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -239,21 +329,45 @@ func TestBulkHappyFlows13Ops(t *testing.T) {
 	cli_initiator.StartDKG.Version = version
 	cli_verify.Verify.Version = version
 	t.Run("test 13 operators 1 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "1", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77,88,99,100,111,122,133", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "1",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77,88,99,100,111,122,133",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
 		resetFlags(RootCmd)
 	})
 	t.Run("test 13 operators 10 validators bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "10", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77,88,99,100,111,122,133", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "10",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77,88,99,100,111,122,133",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
 		resetFlags(RootCmd)
 	})
 	t.Run("test 13 operators 100 validator bulk happy flow", func(t *testing.T) {
-		args := []string{"init", "--validators", "100", "--operatorsInfo", string(operators), "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--operatorIDs", "11,22,33,44,55,66,77,88,99,100,111,122,133", "--nonce", "1", "--amount", "32000000000", "--clientCACertPath", "./certs/rootCA.crt"}
+		args := []string{"init",
+			"--validators", "100",
+			"--operatorsInfo", string(operators),
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--operatorIDs", "11,22,33,44,55,66,77,88,99,100,111,122,133",
+			"--nonce", "1",
+			"--amount", "32000000000",
+			"--clientCACertPath", rootCert[0]}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
@@ -264,7 +378,13 @@ func TestBulkHappyFlows13Ops(t *testing.T) {
 	require.NoError(t, err)
 	validators := []int{1, 10, 100}
 	for i, c := range initCeremonies {
-		args := []string{"verify", "--ceremonyDir", "./output/" + c.Name(), "--validators", strconv.Itoa(validators[i]), "--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494", "--nonce", strconv.Itoa(1), "--amount", "32000000000"}
+		args := []string{"verify",
+			"--ceremonyDir", "./output/" + c.Name(),
+			"--validators", strconv.Itoa(validators[i]),
+			"--withdrawAddress", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--owner", "0x81592c3de184a3e2c0dcb5a261bc107bfa91f494",
+			"--nonce", strconv.Itoa(1),
+			"--amount", "32000000000"}
 		RootCmd.SetArgs(args)
 		err := RootCmd.Execute()
 		require.NoError(t, err)
