+++ integration_test/reshare_bulk_test.go
@@ -109,7 +109,8 @@ func TestBulkReshareHappyFlows4Ops(t *testing.T) {
 				"--newOperatorIDs", "55,66,77,88",
 				"--nonce", "10",
 				"--amount", "32000000000",
-				"--signatures", signature}
+				"--signatures", signature,
+				"--clientCACertPath", rootCert[0]}
 			RootCmd.SetArgs(args)
 			err = RootCmd.Execute()
 			require.NoError(t, err)
@@ -236,7 +237,8 @@ func TestBulkReshareHappyFlows7Ops(t *testing.T) {
 				"--newOperatorIDs", "77,88,99,110,111,112,113",
 				"--nonce", "10",
 				"--amount", "32000000000",
-				"--signatures", signature}
+				"--signatures", signature,
+				"--clientCACertPath", rootCert[0]}
 			RootCmd.SetArgs(args)
 			err = RootCmd.Execute()
 			require.NoError(t, err)
@@ -364,7 +366,8 @@ func TestBulkReshareHappyFlows10Ops(t *testing.T) {
 				"--newOperatorIDs", "11,22,33,44",
 				"--nonce", "10",
 				"--amount", "32000000000",
-				"--signatures", signature}
+				"--signatures", signature,
+				"--clientCACertPath", rootCert[0]}
 			RootCmd.SetArgs(args)
 			err = RootCmd.Execute()
 			require.NoError(t, err)
@@ -492,7 +495,8 @@ func TestBulkReshareHappyFlows13Ops(t *testing.T) {
 				"--newOperatorIDs", "11,22,33,44",
 				"--nonce", "10",
 				"--amount", "2048000000000",
-				"--signatures", signature}
+				"--signatures", signature,
+				"--clientCACertPath", rootCert[0]}
 			RootCmd.SetArgs(args)
 			err = RootCmd.Execute()
 			require.NoError(t, err)
